# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/devices/": {
            "get": {
                "description": "get mobile devices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "List mobile devices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "os",
                        "name": "os",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/device.RentingDeviceResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apperror.HttpError"
                        }
                    }
                }
            }
        },
        "/api/v1/devices/rent/{device_id}": {
            "get": {
                "description": "rent device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Rent device",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Device ID",
                        "name": "device_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/device.RentingDeviceResponse"
                        }
                    },
                    "400": {
                        "description": "User input error, see error detail",
                        "schema": {
                            "$ref": "#/definitions/apperror.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apperror.HttpError"
                        }
                    }
                }
            }
        },
        "/api/v1/devices/return/{device_id}": {
            "get": {
                "description": "return device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Return device",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Device ID",
                        "name": "device_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/device.RentingDeviceResponse"
                        }
                    },
                    "400": {
                        "description": "User input error, see error detail",
                        "schema": {
                            "$ref": "#/definitions/apperror.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apperror.HttpError"
                        }
                    }
                }
            }
        },
        "/users/employee/{empl-id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get users with certain employee_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get users with certain employee_id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "empl-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/user.UserWithProjectsDTO"
                            }
                        }
                    }
                }
            }
        },
        "/users/employees": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get employees list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get employees list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/user.EmpolyeeDTO"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apperror.HttpError": {
            "description": "Information about the error",
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                }
            }
        },
        "device.RentingDeviceResponse": {
            "description": "Information about the current status of device",
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "Aydar Ibragimov"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Iphone 12 Pro"
                }
            }
        },
        "user.EmpolyeeDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "user.UserWithProjectsDTO": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "projects": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}